# Conflicts:
#	paimon-common/src/main/java/org/apache/paimon/globalindex/SortedFileGlobalIndexReader.java
#	paimon-common/src/main/java/org/apache/paimon/globalindex/btree/BTreeIndexReader.java
#	paimon-common/src/main/java/org/apache/paimon/globalindex/btree/LazyFilteredBTreeReader.java
#	paimon-common/src/test/java/org/apache/paimon/globalindex/btree/BTreeIndexReaderTest.java
#	paimon-core/src/main/java/org/apache/paimon/table/source/AbstractBatchTableScan.java
#	paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/source/assigners/DynamicPartitionPruningAssigner.java
#	paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/source/assigners/PreAssignSplitAssigner.java
#	paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/utils/TableScanUtils.java

+++ paimon-common/src/main/java/org/apache/paimon/reader/LimitRecordReader.java
@@ -40,7 +40,7 @@ public final class LimitRecordReader<T> implements RecordReader<T> {
         this.limit = limit;
     }
 
-    public static <T> RecordReader<T> limit(RecordReader<T> reader, @Nullable Integer limit) {
+    public static <T> RecordReader<T> limit(RecordReader<T> reader, @Nullable Long limit) {
         if (limit == null || limit <= 0) {
             return reader;
         }
