+++ paimon-core/src/main/java/org/apache/paimon/table/source/KeyValueTableRead.java
@@ -68,7 +68,7 @@ public final class KeyValueTableRead extends AbstractDataTableRead {
     private Predicate predicate = null;
     private IOManager ioManager = null;
     @Nullable private TopN topN = null;
-    @Nullable private Integer limit = null;
+    @Nullable private Long limit = null;
     @Nullable private ReadBatchSizer readBatchSizer;
 
     public KeyValueTableRead(
@@ -146,7 +146,7 @@ public final class KeyValueTableRead extends AbstractDataTableRead {
     }
 
     @Override
-    public InnerTableRead withLimit(int limit) {
+    public InnerTableRead withLimit(long limit) {
         this.limit = limit;
         return this;
     }
